core/local: Move errors to dedicated module
  Node v14.16.0, coming with Electron v12.0.0, detects a circular
  dependency with our `core/local/index.js` module when importing its
  error definitions in other modules.

  The simple way to fix this issue is to extract those error definitions
  into a separate module and import them wherever they're needed.
