+++ core/local/errors.js
@@ -1,0 +1,12 @@
+/**
+ * @module core/local/errors
+ * @flow
+ */
+
+const SYNC_DIR_EMPTY_MESSAGE = 'Syncdir is empty'
+const SYNC_DIR_UNLINKED_MESSAGE = 'Syncdir has been unlinked'
+
+module.exports = {
+  SYNC_DIR_EMPTY_MESSAGE,
+  SYNC_DIR_UNLINKED_MESSAGE
+}
