+++ core/local/index.js
@@ -55,9 +55,6 @@ export type LocalOptions = {
 }
 */
 
-const SYNC_DIR_EMPTY_MESSAGE = 'Syncdir is empty'
-const SYNC_DIR_UNLINKED_MESSAGE = 'Syncdir has been unlinked'
-
 /** `Local` is the class that interfaces cozy-desktop with the local filesystem.
  *
  * It uses a watcher, based on chokidar, to listen for file and folder changes.
@@ -488,7 +485,5 @@ class Local /*:: implements Reader, Writer */ {
 }
 
 module.exports = {
-  SYNC_DIR_EMPTY_MESSAGE,
-  SYNC_DIR_UNLINKED_MESSAGE,
   Local
 }
